Support non-string header values that stringify
To avoid leaking secrets in stack traces, my team has started passing
secrets as blessed coderefs, similar to Secret.pm in this commit. One of
the spots we have a secret is in a header value.

I found that passing such an object to HTTP::Headers sort of works
already, but when LWP tries to use it in the request it calls clone()
which fails due to Storable throwing "Can't store CODE items".

Instead, this changes us to use Clone::clone which seems to provide
similar functionality while supporting this use case.