only install recommended modules on new enough Perls
We need to do this because IO::Compress::Brotli requires Perl 5.14, but
we don't want to depend on that. So we've made it a recommended
dependency, which we still want to install to run tests with on all
Perls that allow it. But we need to explicitly not attempt to install it
on Perls that are too old, because cpm cannot determine that a module's
dependencies aren't met and ignore it.