Fix version numbers after cherry-picking
v6.41 was accidentally released from a branch and the commits weren't
merged back into master. In the meantime, more commits have gone into
master for v6.42. I've cherry-picked v6.41 into master and now I'm
fixing up the artifacts. The history is going to look a bit wonky in the
repo, but that's probably better than a force-push at this point.